# Conflicts:
#	bsp/boards/openmote-b/radio_subghz.c
#	openstack/02b-MAChigh/sixtop.c

+++ bsp/boards/uart.h
@@ -39,7 +39,7 @@ void    uart_clearRxInterrupts(void);
 void    uart_clearTxInterrupts(void);
 void    uart_writeByte(uint8_t byteToWrite);
 #ifdef FASTSIM
-void    uart_writeCircularBuffer_FASTSIM(uint8_t* buffer, uint8_t* outputBufIdxR, uint8_t* outputBufIdxW);
+void    uart_writeCircularBuffer_FASTSIM(uint8_t* buffer, uint16_t* outputBufIdxR, uint16_t* outputBufIdxW);
 #endif
 uint8_t uart_readByte(void);
